Synthetic dataset generation
Using Unity Perception Package a simulation of the J.J.Pickle Research
Campus was implemented. A randomized dataset was generated by Unity. The
synthetic dataset was uploaded to `roboflow.com` to further augment the
image diversity. Finally, the dataset was exported into YoloV7 format
and a model was trained using the `make train` command.