Detector trained on JJPickle's live data
The video was sliced into approximately 1000 images.
Then, Antonio, Aditi, Aditya and me labeled the images using Roboflow
Finally, the dataset was augmented and exported from Roboflow.

Local training using the `make train` command was performed. It took
around 15 hours.

The result is that the detector works pretty well on the video feed
(~90% confidence) on both the top and side logos. NOTE: this is expected
since the training data came from the video. Therefore, more tests must
be run to conclude anything about the robustness of the solution.

The suggested game plan is to get more images with a lot of diversity
and get videos of the actual setup (crate, final logo design and
environment) where the demonstration is going to take place.